LockBits: acceleration by implementing a fast path for same-buffer image copy operations
the current/old design is this:

for (pixel 0...999)
{
    get_pixel()
    put_pixel()
}

the updated design is this:

if (optimisation_is_applicable)
{
    memcpy(dst, src, 999*pixel_size)
}
else
{
    /* use current/old design */
}

as you can see the current/old code is doing two function calls per loop cycle.
this is definitely a performance kill on even the best application processor.
its hard to say how much worser it will hit the performance on sub-optimal
processor designs.

in get_pixel and put_pixel there is further a bunch of if/else sequences hidden
that is much worser than the one-time check (optimisation_is_applicable).