Enable autotools-based testing
Automakes provides facilities to run tests and collect their
results. This can be enabled simply by adding a TESTS setting which
contains the names of the test programs to be run.

All of the tests in the "tests/" folder are now enabled except for
testgdi, which requires interaction with the user.

Commit licensed under MIT/X11.