Fix testbits
The testbits test failure indicates an issue in the definition and use
of the BitmapData structure. Document the issue and work around it in
the test, to allow the test to actually check the behaviour of the
implementation.

Also see https://bugzilla.xamarin.com/show_bug.cgi?id=20682

Commit licensed under MIT/X11.