Visual C++ support: Define API visibility
When compiling libgdiplus with Visual C++, all functions would be marked as internal because they lack the __declspec( dllexport ) attribute.

To work around this, the WINGDIPAPI is defined symbol (the same symbol which is used in the public GDI+ API, see MSDN), which is left empty when consuming the public API and when on non-msvc compilers, and set to __declspec( dllexport ) when compiling with msvc.