Fix the width and height of EMF images (#113)
* Move Metafile enums to gdipenums.h to match GDI+

* Match Windows GDI+ with definition of MetafileHeader

* Fix width and height of EMF images

This is the actual bug fix to the previously disabled test. If you look at the [MSDN docs](https://msdn.microsoft.com/en-us/library/windows/desktop/ms534065(v=vs.85).aspx)

> **szlDevice**
>     Type: **SIZEL**
>     Resolution, in pixels, of the reference device.

So basically the problem here was that the width and height of the EMF file would actually be the resolution of my screen (1440x900 because I created the file on my laptop).

This is the wrong parameter to use for width and height. I just combined these two cases into the same one, because the `x` and `y` value would be `0`, so subtracting`0` would not change the value.