Remove second translation for metafile in GdipDrawImageRect
GdipDrawImageRect has been changed so that it doesn't translate a metafile a second time.
Previously it drew it at (x, y) and then translated it by (x, y), so the translation has been removed.

This was done because in 000da567cf8669c84fef787dbe3718663f8fc227
gdip_metafile_play_setup() was changed to do translation as well.