Fix WMF dimensions if inches is zero (#190)
The units of a metafile are twips (1/20 of a point). The Inch field contains the number of twips per inch used to represent the image. There are 1440 twips per inch by default.
Use the default if the placeable header has no twips per inch specified.

This aligns us with GDI+ behaviour and fixes #179