Use our own pango font maps to support custom fonts and use FontConfig backend on macOS by default. (#258)
Fixes #237.

Pango has three backends - CoreText, Win32 and FontConfig. By default it tries to select the one that is most appropriate for the platform. However the current libgdiplus code depends on FontConfig and FreeType directly, so we need to force the use of FontConfig backend. Additionally it fixes an issue with custom fonts, which use custom FcConfig objects that were not properly propagated to Pango. Lastly, it fixes a memory leak in gdip_get_fontfamily_details.