Support giflib on Windows (#196)
Fixes #194

The problem was that we crashed calling DGifCloseFile when freeing RasterBits.

Nothing is wrong with the dll though. It's that we're always using the `MultiThreadedDLL` debug library, rather than the `MultiThreadedDebugDLL` runtime library in libgdiplus. Giflib, however, uses the correct runtime library on Debug mode.

However, there's a mismatch between the two which causes this error!