Fix Pango LineLimit cliping.
Check the line fits in with > instead of >=

This is actually not as obvious as it might seem, because clipy2 is
declared as double. It would be reasonable to expect the comparison to
never be equal. However, in a variety of situations clipy2 is set
to FrameHeight, which is an integer. It is being compared to y1
divided by PANGO_SCALE, both of which are integers, resulting in the
possibility of comparing an integer to a double containing an exact
representation of an integer, possibly of the same value.