In Pango AddPathString, don't include a final move
If the Cairo context has a current point set when we call layout_path,
Pango will move back to where it started. That would produce a move
instruction in the path we copy, which we don't want. If the current
point is set after generating the Cairo path, clear it.

The main consequence of the old / incorrect behaviour was that getting
the path bounds would return an incorrect value (the useless extra point
was included). Also layoutRect was always specified when using from
Mono, as GraphicsPath.AddString requires either a Point or Rectangle.