Fix pango_MeasureString chars fitted with multi-byte chars.
Previously the maximum length to measure was tested with lastIndex + 1.
The problem was that if lastIndex happens to be at a multi-byte
character, the + 1 will only allow g_utf8_strlen to measure to part-way
through that character. This means the last character is not counted.

Also do similarly when moving back when lastIndex >= strlen(layoutText).
I'm not sure if this is really needed or not, but there are multi-byte
line endings / new lines that may require it.