GraphicsPath fixes (#383)
* Fix CLOSE_PATH handling. Use a Cairo Recording surface.

The Recording surface is much better than an Image surface because the
Image surface is of limited size. This only affects paths significantly
larger than the surface, but is still an issue.

* Correct recording surface content type.

I hadn't realised recoding surfaces had a different format / content
than image surfaces, and hadn't seen any compiler warnings or other
errors until the Travis CI Mac build failed.

* In Pango AddPathString, don't include a final move

If the Cairo context has a current point set when we call layout_path,
Pango will move back to where it started. That would produce a move
instruction in the path we copy, which we don't want. If the current
point is set after generating the Cairo path, clear it.

The main consequence of the old / incorrect behaviour was that getting
the path bounds would return an incorrect value (the useless extra point
was included). Also layoutRect was always specified when using from
Mono, as GraphicsPath.AddString requires either a Point or Rectangle.

* Fix a positioning bug in GdipAddPathString.

* Add bounds tests for GdipAddPathString.

Also adds a test helper method, for comparing floats
to a specified tolerance.

* Fix a bug in my new test helper method.

* Only run the bounds test when using suitable text rendering.

Cairo doesn't perform line alignment, so this fails there.

* Move test variables inside #ifdef to avoid warnings.