Pango fixes (#384)
* Fix font family collection segfault.

GdipGetFontCollectionList did not set GpFontFamily.collection for the
returned families, resulting in a segmentation fault when getting font
metrics with the Pango text rendering backend.

fontfamily_init did not initialise the family's collection either. This
made the actual problem harder to find, as collection was pointing to
invalid memory (collection->config was pointing at a font family).

* Pango font description needs the size in points.

Pass pango_font_description_set_size the size in points, not whatever
unit font->emSize happens to be using.

* Pango MeasureString measuring whitespace

Measure the actual space passed in, not just a hard-coded normal space
char. The actual whitespace might be something else (e.g. em space).

* Fix Pango LineLimit cliping.

Check the line fits in with > instead of >=

This is actually not as obvious as it might seem, because clipy2 is
declared as double. It would be reasonable to expect the comparison to
never be equal. However, in a variety of situations clipy2 is set
to FrameHeight, which is an integer. It is being compared to y1
divided by PANGO_SCALE, both of which are integers, resulting in the
possibility of comparing an integer to a double containing an exact
representation of an integer, possibly of the same value.

* LineLimit should not be affected by clip

Unless NoClip is set, clipy2 is the lesser of the clip-region-based
height, and the specified height. On Windows, LineLimit is not affected
by the clip region irrespective of NoClip. Changed to always compare the
bottom of the line with FrameHeight.

* Test MeasureString on whitespace-only lines.

The behaviour tested for (measure the first / only character despite
trimming) is the expected behaviour throughout libgdiplus and Mono. I do
not guarantee it matches Windows.

* Test MeasureString LineLimit behaviour.

* Correct text-pango counting of multi-byte characters removed.

Looks like the previous code relied on the number of characters removed
being equal to the difference in length between the input and output.
This does not work correctly with multi-byte characters, as the length
difference is in bytes. Instead count the characters being removed.

* Fix pango_MeasureString chars fitted with multi-byte chars.

Previously the maximum length to measure was tested with lastIndex + 1.
The problem was that if lastIndex happens to be at a multi-byte
character, the + 1 will only allow g_utf8_strlen to measure to part-way
through that character. This means the last character is not counted.

Also do similarly when moving back when lastIndex >= strlen(layoutText).
I'm not sure if this is really needed or not, but there are multi-byte
line endings / new lines that may require it.

* Fix mistake in couning characters removed.

Forgot to decrement the number of characters removed after adding back
the first one on a whitespace-only string. I had it there, but
accidentally removed it when excluding something else.

* Only limit Pango layout to the bounding box.

Limiting the layout to the clip region is incorrect, and upsets the
reported bounds as well as lines and characters fitted.

* Fix MeasureSring expected line count when no lines fitted.

I hadn't bothered to check the correct behaviour and had accepted the
existing. But if LineLimit prevents fitting any lines, the number of
lines fitted should be zero.

* In MeasureString, don't count lines eliminated by LineLimit.

* Fix out-of-bounds read in MeasureString getting chars fitted.

When no characters were fitted, g_utf8_prev_char would go off the start
of layoutText. This also made lastIndex negative, which was then used to
access charsRemoved. Both reads are incorrect.

Fixed by handling the case of nothing fitted separately. In that
situation charsRemoved is not relevant.

* Fix another out-of-bounds read in pango_MeasureString.

When layoutText is of zero length, after my last fix lastIndex was being
left at 0. This meant checking for a line ending (with
layoutText[lastIndex + 1]) was off the end of the string.

* Tweak test parameters to suit a range of fonts.