Don't implement GdipCreateBitmapFromHICON on Windows (#559)
Similar to #557, assuming a `HICON` is a `GpImage` will very likely not work on Windows; so just return `NotImplemented` instead of risking taking down the entire process.

Similar to #557, the corefx unit tests will call this with HICON set to a value of `0xa` (i.e. an invalid pointer), not sure how we can validate that pointer short of tracking all `GpGraphics` objects.