Don't attempt to implement GdipCreateFromHDC on Windows (#557)
The first thing `GdipCreateFromHdc` does, is to cast this from `HDC` to `GpGraphics`.  This is almost guaranteed to fail on Windows, and even something as simple as `clone->type` will cause an access violation.

Instead, just return `NotImplemented` when compiled for Win32.

A more robust strategy may be to keep track of all `GpGraphics` objects we allocate, and check whether `hdc` points to any of those, before casting `hdc` to `GpGraphics`.