GdipCreateBitmapFromScan0: Fix a buffer overflow with pixel formats < 32 bpp (#558)
If you create a new bitmap from scratch (that is, `scan0` is `NULL`), `GdipCreateBitmapFromScan0` will allocate memory to hold the raw bitmap data for you.

It then attempts to initialize this. By default, it will `memset` all bytes to 0, so you should get a black image in most formats. However, for formats with alpha-channels, this would set alpha to `0`, causing the image to be "completely translucent black" (AKA white).

So the method goes on a quest to set the alpha channel to 1. It iterates over the framebuffer using a 32-bit `ARGB` value. For pixel formats with < 32 bpp, this would cause a buffer overflow for the last pixel.

Fix the check for the bpp, and make sure we really have a 32bpp pixel format.