GdipGetRegionScans: Operator on work instead of region (#556)
`GdipGetRegionScans` transforms `region` and stores the result in `work`.  From that point on, all work is done on `work` and no longer on region.

However, there still was a switch on `region->type` instead of `work->type` which would send `GdipGetRegionScans` down the wrong path if the type has changed.

To make things worse, `gdip_region_convert_to_path` changes the `type` and clears `rects`, but did not reset `cnt` to 0.

`RegionTests.GetRegionScans_CustomMatrix_TransformsRegionScans` in the corefx test suite could trigger this.