Bump minimum version of Pango and add C version of MeasureString_Wrapping_Dots test (#574)
* Add C version of MeasureString_Wrapping_Dots test

* Bump minimum Pango version to 1.40.14 to get fix for line breaking bug

This is one funky bug... It works fine on macOS with Pango 1.42 and it doesn't work on Ubuntu 16.04 with Pango 1.38.1. The problem is that on Ubuntu the "." characters are not marked with "char_break" logical attribute so the Pango layout fails to insert a line break. On macOS all the characters have the "char_break" logical attribute and it returns the correct result.

The fix in Pango was implemented in GNOME/pango@e8316c1#diff-3d8168315b940c3e9d29a4991735a9a7 so any version newer than 1.40.14 is fine.