Fix multiple bugs and conversions for PARGB32 bitmap object and premultiplied Cairo surfaces (#599)
* Convert to/from premultiplied colors in GdipBitmap[Set/Get]Pixel in PARGB32 mode

* Unify all the places where Cairo surface is created from raw GpBitmap data. If necessary, create a
copy of the original data and premultiply the RGB channels with the alpha value to ensure that the
Cairo surface operates on the PARGB32 format it expects. When accessing the GpBitmap data and a
Cairo surface exists ensure that we either use the data from the surface (GdipBitmapGetPixel,
GdipBitmapSetPixel) or that we update the raw data from the surface before accessing them. This
only applies to conversions between ARGB32 and PARGB32 surfaces, all other ones use the original
buffers directly. Also, when creating GpGraphics object from GpBitmap using
GdipGetImageGraphicsContext reuse any pre-existing surface that is already associated with
GpBitmap.

Contributes to #446.

Fixes the following issues:
- Converts to/from premultiplied colors in GdipBitmap[Set/Get]Pixel in PARGB32 mode
- Converts back and forth underlying data for Cairo surfaces where only PARGB32 is supported and ARGB32 is emulated. This fixes regressions caused by 389cb7bb49006c0a598ee3369d9c4876574740fe and f7775ec5746d242e9c3c0f06a725bfee748f8243 at the expense of slight inaccuracies due to rounding in the ARGB32 -> PARGB32 -> ARGB32 conversion.