Use surface instead of scan0 while image rotation if it is not empty (#614)
While image rotating, surface existance and premultiplication format should be checked. In first case, surface should be usea as a source for rotting. In other case - old behaviour. See: bitmap.c@GdipBitmapGetPixel
In the end of function old surface should be replaced to new one.

This opened task described problem and the way to reproduce:
https://github.com/mono/mono/issues/17743