Fix text baseline position for some fonts
Some fonts (particularly Calibri or the metric-equivalent Carlito) were
not aligned to the same baseline as on Windows. This is because fonts
have three sets of roughly equivalent metrics, and different platforms
use different ones. Pango uses one set, and Windows uses a different
set. The one Windows uses has the line spacing included in the ascent
and descent. The other two allow the gap to be specified separately,
but the one Pango seems to use has normally got only a minimal gap.

The problem is that Pango is aligning the top of the font
according to just the ascent, and on fonts like the above where this is
considerably different to the Windows ascent the text is positioned
equivalently higher in the bounding box.