Fixes
Don't adjust baseline on vertical text. It needs to be handled
differently to horizontal, and has enough other problems that I don't
want to start. Fixes test failure.

Fixes the compile error that happened for Windows.

Report the baseline offset separately to the box offset, because it
needs to be handled differently for measurement etc. Basically when
drawing it needs to make it draw lower (affecting position), but when
measuring it needs to make the rectangle taller (affecting size but
leaving the position unchanged).