Allow GpGraphics to take ownership of GpMetafile on dispose
The Windows GDI+ allows the metafile instance to be disposed before the
graphics instance.

To support that, keep track in the GpGraphics of whether it is responsible for
freeing the GpMetafile.

If the graphics instance is deleted first, own_metafile is FALSE and nothing
happens.

If the metafile instance is disposed first, set own_metafile to TRUE and don't
dispose of the metafile. It will be disposed when the graphics instance is
deleted.

---

Also on Windows GDI+ creating more than one graphics instance for a metafile
instance throws an OutOfMemoryException, so do the same in GdipGetImageGraphicsContext